// apps/api/src/common/enums/roles.enum.ts
export enum UserRole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADL_ADMIN = 'ADL_ADMIN',
  SCHOOL_GATEKEEPER = 'SCHOOL_GATEKEEPER',
  TRAINER = 'TRAINER',
  STUDENT = 'STUDENT',
  PARENT = 'PARENT',
  ACCOUNTANT = 'ACCOUNTANT',
  SYSTEM_AUDITOR = 'SYSTEM_AUDITOR',
}

export const GLOBAL_ROLES = [UserRole.SUPER_ADMIN, UserRole.ADL_ADMIN];
export const SCHOOL_ADMIN_ROLES = [UserRole.SCHOOL_GATEKEEPER];
export const READ_ONLY_ROLES = [UserRole.SYSTEM_AUDITOR];